Moved ErrorBound to common crate
