+++ crates/modelardb_common/src/types.rs
@@ -19,6 +19,10 @@
 //! incorrect aliases if types from different modules use the same name. It is assumed that each set
 //! of aliases are all for the same underlying type.
 
+use std::cmp::Ordering;
+
+use crate::errors::ModelarDbError;
+
 // Types used for a univariate id.
 pub type UnivariateId = std::primitive::u64;
 pub type ArrowUnivariateId = arrow::datatypes::UInt64Type;
@@ -57,3 +61,89 @@ pub struct QuerySchema(pub arrow::datatypes::SchemaRef);
 
 #[derive(Clone)]
 pub struct ConfigurationSchema(pub arrow::datatypes::SchemaRef);
+
+/// Error bound in percentage that is guaranteed to be from 0.0% to 100.0%. For both `PMCMean`,
+/// `Swing`, and `Gorilla` the error bound is interpreted as a relative per value error bound.
+#[derive(Debug, Copy, Clone)]
+pub struct ErrorBound(pub f32); // Simpler for the model types to directly work on the f32.
+
+impl ErrorBound {
+    /// Return [`ErrorBound`] if `percentage` is a value from 0.0% to 100.0%, otherwise
+    /// [`CompressionError`](ModelarDbError::CompressionError) is returned.
+    pub fn try_new(percentage: f32) -> Result<Self, ModelarDbError> {
+        if !(0.0..=100.0).contains(&percentage) {
+            Err(ModelarDbError::CompressionError(
+                "Error bound must be a value from 0.0% to 100.0%.".to_owned(),
+            ))
+        } else {
+            Ok(Self(percentage))
+        }
+    }
+
+    /// Consumes `self`, returning the error bound as a [`f32`].
+    pub fn into_inner(self) -> f32 {
+        self.0
+    }
+}
+
+/// Enable equal and not equal for [`ErrorBound`] and [`f32`].
+impl PartialEq<ErrorBound> for f32 {
+    fn eq(&self, other: &ErrorBound) -> bool {
+        self.eq(&other.0)
+    }
+}
+
+/// Enable less than and greater than for [`ErrorBound`] and [`f32`].
+impl PartialOrd<ErrorBound> for f32 {
+    fn partial_cmp(&self, other: &ErrorBound) -> Option<Ordering> {
+        self.partial_cmp(&other.0)
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use arrow::array::BinaryArray;
+    use modelardb_common::types::{TimestampArray, ValueArray};
+    use modelardb_common_test::data_generation::{self, ValuesStructure};
+    use proptest::num;
+    use proptest::proptest;
+
+    use crate::compression;
+
+    const ERROR_BOUND_ZERO: f32 = 0.0;
+    const UNCOMPRESSED_TIMESTAMPS: &[Timestamp] = &[100, 200, 300, 400, 500];
+
+    // Tests for ErrorBound.
+    proptest! {
+    #[test]
+    fn test_error_bound_can_be_positive_if_less_than_one_hundred(percentage in num::f32::POSITIVE) {
+        if percentage <= 100.0 {
+            assert!(ErrorBound::try_new(percentage).is_ok())
+        } else {
+            assert!(ErrorBound::try_new(percentage).is_err())
+        }
+    }
+
+    #[test]
+    fn test_error_bound_cannot_be_negative(percentage in num::f32::NEGATIVE) {
+        assert!(ErrorBound::try_new(percentage).is_err())
+    }
+    }
+
+    #[test]
+    fn test_error_bound_cannot_be_positive_infinity() {
+        assert!(ErrorBound::try_new(f32::INFINITY).is_err())
+    }
+
+    #[test]
+    fn test_error_bound_cannot_be_negative_infinity() {
+        assert!(ErrorBound::try_new(f32::NEG_INFINITY).is_err())
+    }
+
+    #[test]
+    fn test_error_bound_cannot_be_nan() {
+        assert!(ErrorBound::try_new(f32::NAN).is_err())
+    }
+}
