+++ crates/modelardb_compression_python/src/lib.rs
@@ -24,8 +24,8 @@ use arrow::datatypes::Schema;
 use arrow::error::ArrowError;
 use arrow::record_batch::RecordBatch;
 use modelardb_common::schemas::{COMPRESSED_SCHEMA, QUERY_SCHEMA, UNCOMPRESSED_SCHEMA};
-use modelardb_common::types::{TimestampArray, TimestampBuilder, ValueArray, ValueBuilder};
-use modelardb_compression::{self, ErrorBound};
+use modelardb_common::types::{ErrorBound, TimestampArray, TimestampBuilder, ValueArray, ValueBuilder};
+use modelardb_compression::{self};
 use pyo3::import_exception;
 use pyo3::prelude::{pyfunction, pymodule, wrap_pyfunction, PyAny, PyModule, PyResult, Python};
