+++ crates/modelardb_server/src/metadata/mod.rs
@@ -40,8 +40,7 @@ use datafusion::execution::options::ParquetReadOptions;
 use futures::TryStreamExt;
 use log::LevelFilter;
 use modelardb_common::errors::ModelarDbError;
-use modelardb_common::types::{Timestamp, UnivariateId, Value};
-use modelardb_compression::ErrorBound;
+use modelardb_common::types::{ErrorBound, Timestamp, UnivariateId, Value};
 use sqlx::database::HasArguments;
 use sqlx::error::Error;
 use sqlx::query::Query;
