+++ crates/modelardb_server/src/metadata/model_table_metadata.rs
@@ -21,8 +21,7 @@ use std::sync::Arc;
 use datafusion::arrow::datatypes::{ArrowPrimitiveType, DataType, Schema};
 use datafusion::logical_expr::expr::Expr;
 use modelardb_common::errors::ModelarDbError;
-use modelardb_common::types::{ArrowTimestamp, ArrowValue};
-use modelardb_compression::ErrorBound;
+use modelardb_common::types::{ArrowTimestamp, ArrowValue, ErrorBound};
 
 /// Metadata required to ingest data into a model table and query a model table.
 #[derive(Debug, Clone)]
@@ -181,7 +180,7 @@ mod test {
     use super::*;
 
     use datafusion::arrow::datatypes::{DataType, Field, Schema};
-    use modelardb_common::types::{ArrowTimestamp, ArrowValue};
+    use modelardb_common::types::{ArrowTimestamp, ArrowValue, ErrorBound};
 
     // Tests for ModelTableMetadata.
     #[test]
