+++ crates/modelardb_server/src/parser.rs
@@ -29,8 +29,7 @@ use datafusion::logical_expr::{AggregateUDF, Expr as DFExpr, ScalarUDF, TableSou
 use datafusion::physical_expr::planner;
 use datafusion::sql::planner::{ContextProvider, PlannerContext, SqlToRel};
 use datafusion::sql::TableReference;
-use modelardb_common::types::{ArrowTimestamp, ArrowValue};
-use modelardb_compression::ErrorBound;
+use modelardb_common::types::{ArrowTimestamp, ArrowValue, ErrorBound};
 use sqlparser::ast::{
     ColumnDef, ColumnOption, ColumnOptionDef, DataType as SQLDataType, GeneratedAs,
     HiveDistributionStyle, HiveFormat, Ident, ObjectName, Statement, TimezoneInfo,
